fix: [layout] Vuelta a agregar KEYBOARD_12KEY para ingreso de Slope. fix: Description/Block Dimention: debe aceptar decimales. fix: NumberOfJoints se estaba escribiendo con el valor de Orientation.

+++ Module/src/main/java/com/metric/skava/data/dao/impl/dropbox/helper/AssessmentBuilder4DropBox.java
@@ -83,8 +83,8 @@ public class AssessmentBuilder4DropBox {
         FractureType fractureType = fractureTypeDAO.getFractureTypeByCode(fractureTypeID);
         babyAssessment.setFractureType(fractureType);
 
-        Long blockSize = assessmentRecord.getLong("blockSize");
-        babyAssessment.setBlockSize(blockSize.shortValue());
+        Double blockSize = assessmentRecord.getDouble("blockSize");
+        babyAssessment.setBlockSize(blockSize.doubleValue());
 
         Long numJoints = assessmentRecord.getLong("numJoints");
         babyAssessment.setNumberOfJoints(numJoints.shortValue());
