+++ Module/src/main/java/com/metric/skava/report/fragment/MappingReportMainFragment.java
@@ -149,7 +149,7 @@ public class MappingReportMainFragment extends SkavaFragment {
             ((TextView) rootView.findViewById(R.id.report_tipo_fracturacion_value)).setText(fractureType.getName());
         }
 
-        Short blockSize = currentAssessment.getBlockSize();
+        Double blockSize = currentAssessment.getBlockSize();
         if (blockSize != null) {
             ((TextView) rootView.findViewById(R.id.report_medida_bloques_value)).setText(numberFormat.format(blockSize));
         }
